octdata-crawlers-all-in/diario-da-manha/diario-da-manha/requirements.txt
bs4
requests
sshtunnel
paramiko==3.0
pymongo
pyaml